+++ pages/_app.tsx
@@ -1,8 +1,16 @@
+import { useState } from 'react'
+import { Context } from '../Context'
 import '../styles/globals.css'
 import type { AppProps } from 'next/app'
 
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  const [context, setContext] = useState({ userId: 0, token: '', loggedIn: false });
+
+  return (
+    <Context.Provider value={[ context, setContext ]}>
+      <Component {...pageProps} />
+    </Context.Provider>
+  )
 }
 
 export default MyApp