+++ CHANGELOG.md
@@ -17,3 +17,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - BookInfo component
 - Authors page
 - AuthorInfo component
+- Login, Signin, logout functionality